feat(handlers): complete handler refactoring and issue #41 fixes
Break down 1,862-line migration_handler.go into focused modules:
- shared.go: Constants and types (157 lines)
- base_handler.go: Common utilities (118 lines)
- user_handler.go: User CRUD operations (320 lines)
- tracker_handler.go: Tracker operations with TrueSkill (359 lines)
- admin_handler.go: Dashboard and API endpoints (263 lines)
- validation.go: Validation logic and security (629 lines)

Issue #41 fixes:
- Add Tracker button always visible on user detail pages
- Discord ID pre-filling from URL parameters
- TrueSkill auto-update after tracker creation

Closes #41
