+++ internal/usl/handlers/validation_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestValidateTracker(t *testing.T) {
 	// Create a handler instance (we don't need real dependencies for validation tests)
-	handler := &MigrationHandler{}
+	baseHandler := &BaseHandler{}
 
 	tests := []struct {
 		name        string
@@ -161,7 +161,7 @@ func TestValidateTracker(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			result := handler.validateTracker(tt.tracker)
+			result := baseHandler.validateTracker(tt.tracker)
 
 			if result.IsValid != tt.expectValid {
 				t.Errorf("Expected valid=%v, got valid=%v", tt.expectValid, result.IsValid)
